Build service and module + move into folder
