+++ packages/twenty-server/src/database/commands/data-seed-demo-workspace/crons/data-seed-demo-workspace.cron.command.ts
@@ -2,8 +2,8 @@ import { Inject } from '@nestjs/common';
 
 import { Command, CommandRunner } from 'nest-commander';
 
-import { dataSeedDemoWorkspaceCronPattern } from 'src/database/commands/crons/data-seed-demo-workspace-cron-pattern';
-import { DataSeedDemoWorkspaceJob } from 'src/database/jobs/data-seed-demo-workspace.job';
+import { dataSeedDemoWorkspaceCronPattern } from 'src/database/commands/data-seed-demo-workspace/crons/data-seed-demo-workspace-cron-pattern';
+import { DataSeedDemoWorkspaceJob } from 'src/database/commands/data-seed-demo-workspace/jobs/data-seed-demo-workspace.job';
 import { MessageQueue } from 'src/integrations/message-queue/message-queue.constants';
 import { MessageQueueService } from 'src/integrations/message-queue/services/message-queue.service';
