+++ packages/twenty-server/src/database/commands/data-seed-demo-workspace/jobs/data-seed-demo-workspace.job.ts
@@ -1,0 +1,16 @@
+import { Injectable } from '@nestjs/common';
+
+import { MessageQueueJob } from 'src/integrations/message-queue/interfaces/message-queue-job.interface';
+
+import { DataSeedDemoWorkspaceService } from 'src/database/commands/data-seed-demo-workspace/services/data-seed-demo-workspace.service';
+
+@Injectable()
+export class DataSeedDemoWorkspaceJob implements MessageQueueJob<undefined> {
+  constructor(
+    private readonly dataSeedDemoWorkspaceService: DataSeedDemoWorkspaceService,
+  ) {}
+
+  async handle(): Promise<void> {
+    await this.dataSeedDemoWorkspaceService.seedDemo();
+  }
+}
